chore: rename Themalijnen to ThemalijnenBlock #112
